Move CriteriaParameter into the criteriaParameters file
